Elenco Contenitori Inserita lettura da database locale e creazione record

+++ lib/widgets/contenitore_frigo_widget.dart
@@ -14,7 +14,7 @@ class ContenitoreFrigoWidget extends StatelessWidget {
   });
 
   final String nomeContenitore;
-  final int quantitaDisponibile;
+  final double quantitaDisponibile;
   final int porzioniDisponibili;
   final DateTime dataInserimento;
